Update SDK POJOs to have Lombok annotations (#463)
moves all java sdk POJOs to lombok annotations.
We will have
@DaTa
@builder
@NoArgsConstructor
@AllArgsConstructor

Our goal is to not break current backward compatibility. We have added unit tests in a previous PR for all the accessors and constructors which are deleted as part of Lombokisation.

This will be a base on which we will sync sdk POJOs with server POJOs.